pyproject.toml
[project]
name = "dither"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numba>=0.61.2",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
]
